Demonstrate odd behavior with unless blocks in before filters
If a before filter redirects, subsequent filters are not run. If
subsequent filters have unless blocks, however, the unless blocks will
be run.

This can be problematic in cases where the unless block
expects state from filters that normally run between the redirecting
filter and the filter with the unless block.